[Improve] RoomService를 Query와 Command로 분리
