+++ src/main/java/com/example/real_chat/api/command/RoomCommandApiController.java
@@ -1,28 +1,23 @@
-package com.example.real_chat.api;
+package com.example.real_chat.api.command;
 
 import com.example.real_chat.dto.common.CommonApiResult;
-import com.example.real_chat.dto.common.Result;
 import com.example.real_chat.dto.room.request.CreateRoomRequestDto;
 import com.example.real_chat.dto.room.request.UpdateRoomRequestDto;
 import com.example.real_chat.dto.room.response.CreateRoomResponseDTO;
-import com.example.real_chat.dto.room.response.RoomResponseDTO;
 import com.example.real_chat.entity.room.ChatRoom;
-import com.example.real_chat.service.RoomService;
+import com.example.real_chat.service.command.RoomCommandService;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @RestController
 @RequiredArgsConstructor
-@RequestMapping(value = "api/v1/room")
-public class RoomApiController {
+@RequestMapping(value = "api/v1/room/command")
+public class RoomCommandApiController {
 
-    private final RoomService roomService;
+    private final RoomCommandService roomService;
 
     @PostMapping()
     public ResponseEntity<CreateRoomResponseDTO> createRoom(
@@ -32,29 +27,6 @@ public class RoomApiController {
         Long roomId = roomService.addRoom(chatRoom);
 
         return ResponseEntity.ok().body(new CreateRoomResponseDTO(roomId));
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<RoomResponseDTO> getRoom(
-            @PathVariable Long id
-    ) {
-        ChatRoom chatRoom = roomService.getRoom(id);
-
-        return ResponseEntity.ok().body(new RoomResponseDTO(id, chatRoom.getName()));
-    }
-
-    @GetMapping()
-    public ResponseEntity<Result<?>> getAllRooms() {
-        List<ChatRoom> chatRooms = roomService.getAllRooms();
-
-        return ResponseEntity.ok().body(new Result<>(chatRooms));
-    }
-
-    @GetMapping("/undeleted")
-    public ResponseEntity<Result<?>> getUnDeleteRooms() {
-        List<ChatRoom> unDeletedRooms = roomService.getUnDeletedRooms();
-
-        return ResponseEntity.ok().body(new Result<>(unDeletedRooms));
     }
 
     @DeleteMapping("/{id}")
