+++ src/main/java/com/example/real_chat/service/command/RoomCommandService.java
@@ -1,4 +1,4 @@
-package com.example.real_chat.service;
+package com.example.real_chat.service.command;
 
 import com.example.real_chat.entity.room.ChatRoom;
 import com.example.real_chat.repository.RoomRepository;
@@ -6,12 +6,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class RoomService {
+public class RoomCommandService {
 
     private final RoomRepository roomRepository;
 
@@ -19,27 +17,14 @@ public class RoomService {
         return roomRepository.save(chatRoom);
     }
 
-    public ChatRoom getRoom(Long roomId) {
-        return roomRepository.findById(roomId).orElseThrow();
+    public void updateChatRoom(Long roomId, String name) {
+        ChatRoom room = roomRepository.findById(roomId).orElseThrow();
+        room.update(name);
     }
 
     public void deleteRoom(Long roomId) {
         ChatRoom chatRoom = roomRepository.findById(roomId).orElseThrow();
         if (chatRoom.isDeleted()) throw new RuntimeException();
         else chatRoom.delete();
-    }
-
-    public List<ChatRoom> getAllRooms() {
-        return roomRepository.findAll()
-                .orElseThrow();
-    }
-
-    public List<ChatRoom> getUnDeletedRooms() {
-        return roomRepository.findUnDeletedRooms().orElseThrow();
-    }
-
-    public void updateChatRoom(Long roomId, String name) {
-        ChatRoom room = roomRepository.findById(roomId).orElseThrow();
-        room.update(name);
     }
 }
